+++ src/test/java/com/jcraft/jsch/AlgorithmsIT.java
@@ -376,7 +376,9 @@ public class AlgorithmsIT {
         "arcfour128,none",
         "arcfour128,zlib@openssh.com",
         "rijndael-cbc@lysator.liu.se,none",
-        "rijndael-cbc@lysator.liu.se,zlib@openssh.com"
+        "rijndael-cbc@lysator.liu.se,zlib@openssh.com",
+        "cast128-cbc,none",
+        "cast128-cbc,zlib@openssh.com"
       })
   public void testCiphers(String cipher, String compression) throws Exception {
     JSch ssh = createRSAIdentity();